refactor(core): move more action and policy logic
